- fixed lead's flaws;
